Fix validation of generic proposal form.
