+++ packages/joy-proposals/src/stories/ProposalForms.stories.tsx
@@ -1,24 +1,29 @@
 import React from 'react';
 import '../index.css';
-import { EditForm } from '../forms/GenericProposalForm';
-import { ProposalId, ProposalType, ProposalFormValues, ProposalGenericProp, ProposalValidationConstraints } from '..s/forms/ProposalTypes';
+import {EditForm} from '../forms/GenericProposalForm';
 
 export default {
 	title: 'Proposals | Forms',
 };
 
 const MockGenericProposal = {
+  id: 123,
 	title: 'Please send me some tokens for coffee',
 	description: 'I am a good guy and deserve this reward.'
 }
 
 const MockProposalConstraints = {
-	title: { min: 1, max: 50 },
-	description: { min: 1, max: 50 }
+	title: { min: 3, max: 50 },
+	description: { min: 10, max: 1000 }
 }
 
-export const GenericForm = () =>
+export const DefaultGenericForm = () =>
+	<EditForm
+		constraints={MockProposalConstraints}
+	/>
+
+export const FilledGenericForm = () =>
 	<EditForm
 		entity={MockGenericProposal}
-		constraints={MockProposalConstraints} 
+		constraints={MockProposalConstraints}
 	/>
